requirements.txt
# Required packages for License Plate Information System
pillow>=9.0.0